move CTRE optimization to module
