+++ src/main/java/frc/robot/subsystems/Module.java
@@ -131,6 +131,34 @@ public class Module extends SubsystemBase {
   }
 
   /**
+  * Minimize the change in heading the desired swerve module state would require by potentially
+  * reversing the direction the wheel spins. Customized from WPILib's version to include placing
+  * in appropriate scope for CTRE onboard control.
+  *
+  * @param desiredState The desired state.
+  * @param currentAngle The current module angle.
+  */
+  public static SwerveModuleState optimize(SwerveModuleState desiredState, Rotation2d currentAngle) {
+    double targetAngle = MathUtil.inputModulus(
+      desiredState.angle.getRadians(), 
+      currentAngle.getRadians() - Math.PI,
+      currentAngle.getRadians() + Math.PI
+    );
+    double targetSpeed = desiredState.speedMetersPerSecond;
+    
+    if (Math.abs(targetAngle - currentAngle.getRadians()) > Math.PI/2) {
+      targetAngle = MathUtil.inputModulus(
+        desiredState.angle.getRadians(), 
+        currentAngle.getRadians() - Math.PI/2,
+        currentAngle.getRadians() + Math.PI/2
+      );
+      targetSpeed = -targetSpeed;
+    }
+
+    return new SwerveModuleState(targetSpeed, new Rotation2d(targetAngle));
+  }
+
+  /**
    * set module desired state
    * @param desiredState state to set module to
    * @param isOpenLoop don't use integrated velocity PID and instead set percentage power for state velocity
